requirements.txt
flask
pyjwt

requests
werkzeug
